requirements.txt
Pillow
rembg
numpy
colorama
pytest